[refactor] bidMain 구조 변경
