+++ client/src/Screen/02_SecondTab/Designer/waitBidListScreen.js
@@ -1,11 +1,32 @@
 import React, { useState, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { StyleSheet, Text, View, ScrollView, Image } from 'react-native';
 
+import { getBidListByDesignerId } from '../../../Contexts/Designer/Bid';
+
+import NoProcessBidList from './noProcessBidList';
 import Line from '../../../Components/Common/line';
+import Loading from '../../../Components/Common/loading';
 import ItemHeader from '../../../Components/ListItem/itemHeader';
 import ItemContent from '../../../Components/ListItem/itemContent';
 
-function WaitBidListScreen({ navigation, bidList }) {
+function WaitBidListScreen({ navigation }) {
+  const dispatch = useDispatch();
+  const { data: userInfo } = useSelector((state) => state.user);
+  const {
+    data: bidList,
+    loading: bidLoading,
+    error: bidError,
+  } = useSelector((state) => state.designerBid);
+
+  useEffect(() => {
+    dispatch(getBidListByDesignerId(userInfo.id));
+  }, [dispatch]);
+  if (bidLoading || bidError || !bidList) return <Loading />;
+
+  if (!bidList.length) {
+    return <NoProcessBidList navigation={navigation} />;
+  }
   return (
     <ScrollView style={styles.container}>
       {bidList.map((bid, index) => (
